Rotation hinzugefügt

+++ core/src/com/programmierbeleg/machine_mayhem/SpielObjekte/Feld.java
@@ -1,7 +1,7 @@
 package com.programmierbeleg.machine_mayhem.SpielObjekte;
 
 public class Feld extends SpielObjekt{
-    public Feld(float x, float y, float breite, float höhe, boolean Sichtbar) {
+    public Feld(float x, float y, int breite, int höhe, boolean Sichtbar) {
         super(x, y, breite, höhe, Sichtbar, "Feld");
     }
